undo: Ensure that each buffer has a starting checkpoint (#159)
With this change, clearing the undo sequence will still ensure that each
buffer starts with an undo checkpoint.

Previously, this was *almost* always the case, but on reload of an
individual file we would clear out the checkpoints for all files, and
only re-checkpoint the file we reloaded. This caused several weird
issues, chief among which was that you cause a file to show as being
changed (along with the action sensitivity issues accompanying this).

(cherry picked from commit 09ecc4188626d48ed25f03287103d33e2082a375)