undo: Change the end_group() behaviour to handle a missing group
This is definitely a programming error, but we shouldn't rely on seeing
assertions. More to the point, this will only ever break the user's
experience, so it's almost certainly better here to just discard the
bad group close and warn about it.

(cherry picked from commit cbe85eda0cd26e86a0bd86cf5415305545863309)